#14 - Created new protos, relocated existing files
