+++ grpc-lib/src/main/proto/events/component/component_events.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 option java_multiple_files = true;
 option java_package = "is.yarr.qilletni.grpc.gen";
-option java_outer_classname = "EventsProto";
+option java_outer_classname = "ComponentEventsProto";
 
 // The greeting service definition.
 service Simple {
