+++ grpc-lib/src/main/proto/events/component/function_events.proto
@@ -1,11 +1,11 @@
 syntax = "proto3";
 
-import "events.proto";
+import "events/component/component_events.proto";
 
 option java_multiple_files = true;
-option java_package = "is.yarr.qilletni.grpc.gen.events.function";
+option java_package = "is.yarr.qilletni.grpc.gen.events.component.function";
 
-service Function {
+service FunctionService {
   rpc Create(FunctionCreateEvent) returns (CreateComponentResponse);
   rpc ChangeChildren(FunctionChangeChildrenEvent) returns (EmptyResponse);
   rpc ChangeName(FunctionNameChangeEvent) returns (EmptyResponse);
