+++ grpc-lib/src/main/proto/events/component/general_events.proto
@@ -1,11 +1,11 @@
 syntax = "proto3";
 
-import "events.proto";
+import "events/component/component_events.proto";
 
 option java_multiple_files = true;
-option java_package = "is.yarr.qilletni.grpc.gen.events.general";
+option java_package = "is.yarr.qilletni.grpc.gen.events.component.general";
 
-service General {
+service GeneralService {
   rpc Delete(DeleteEvent) returns (EmptyResponse);
   rpc Recolor(RecolorEvent) returns (EmptyResponse);
 }
