+++ grpc-lib/src/main/proto/events/component/song_events.proto
@@ -1,11 +1,11 @@
 syntax = "proto3";
 
-import "events.proto";
+import "events/component/component_events.proto";
 
 option java_multiple_files = true;
-option java_package = "is.yarr.qilletni.grpc.gen.events.song";
+option java_package = "is.yarr.qilletni.grpc.gen.events.component.song";
 
-service Song {
+service SongService {
   rpc Create(SongCreateEvent) returns (CreateComponentResponse);
   rpc ChangeSong(SongChangeEvent) returns (EmptyResponse);
 }
