+++ src/main/java/is/yarr/qilletni/grpc/events/FunctionService.java
@@ -6,10 +6,10 @@ import is.yarr.qilletni.components.FunctionComponent;
 import is.yarr.qilletni.database.repositories.components.FunctionComponentRepository;
 import is.yarr.qilletni.grpc.gen.CreateComponentResponse;
 import is.yarr.qilletni.grpc.gen.EmptyResponse;
-import is.yarr.qilletni.grpc.gen.events.function.FunctionChangeChildrenEvent;
-import is.yarr.qilletni.grpc.gen.events.function.FunctionCreateEvent;
-import is.yarr.qilletni.grpc.gen.events.function.FunctionGrpc;
-import is.yarr.qilletni.grpc.gen.events.function.FunctionNameChangeEvent;
+import is.yarr.qilletni.grpc.gen.events.component.function.FunctionChangeChildrenEvent;
+import is.yarr.qilletni.grpc.gen.events.component.function.FunctionCreateEvent;
+import is.yarr.qilletni.grpc.gen.events.component.function.FunctionNameChangeEvent;
+import is.yarr.qilletni.grpc.gen.events.component.function.FunctionServiceGrpc;
 import is.yarr.qilletni.utility.UUIDUtils;
 import net.devh.boot.grpc.server.service.GrpcService;
 import org.springframework.security.access.annotation.Secured;
@@ -19,7 +19,7 @@ import java.util.UUID;
 import static is.yarr.qilletni.grpc.security.Authorities.GENERAL;
 
 @GrpcService
-public class FunctionService extends FunctionGrpc.FunctionImplBase {
+public class FunctionService extends FunctionServiceGrpc.FunctionServiceImplBase {
 
     private final FunctionComponentRepository functionComponentRepository;
     private final GRPCRepositoryInterfacer<FunctionComponent> repositoryInterfacer;
