+++ src/main/java/is/yarr/qilletni/grpc/events/SongService.java
@@ -3,14 +3,14 @@ package is.yarr.qilletni.grpc.events;
 import io.grpc.stub.StreamObserver;
 import is.yarr.qilletni.grpc.gen.CreateComponentResponse;
 import is.yarr.qilletni.grpc.gen.EmptyResponse;
-import is.yarr.qilletni.grpc.gen.events.song.SongChangeEvent;
-import is.yarr.qilletni.grpc.gen.events.song.SongCreateEvent;
-import is.yarr.qilletni.grpc.gen.events.song.SongGrpc;
+import is.yarr.qilletni.grpc.gen.events.component.song.SongChangeEvent;
+import is.yarr.qilletni.grpc.gen.events.component.song.SongCreateEvent;
+import is.yarr.qilletni.grpc.gen.events.component.song.SongServiceGrpc;
 import org.springframework.security.access.annotation.Secured;
 
 import static is.yarr.qilletni.grpc.security.Authorities.GENERAL;
 
-public class SongService extends SongGrpc.SongImplBase {
+public class SongService extends SongServiceGrpc.SongServiceImplBase {
 
     @Override
     @Secured({GENERAL})
